backend/requirements.txt
Flask==2.2.2
Werkzeug<3.0.0
requests
beautifulsoup4
lxml
openai
python-dotenv
